fix(5): refactor post/user logic to PostService
